M 优化打包日志字段类型

+++ model/sql/log/game_packaging_log.go
@@ -9,10 +9,10 @@ import (
 type OdsGamePackagingLogModel struct {
 	sql.SqlBaseModel
 	Db              func() *gorm.DB `json:"-" gorm:"-"`
-	PlatformId      int             `json:"platform_id" gorm:"column:platform_id;default:0;comment:平台ID"`
-	GameId          int             `json:"game_id" gorm:"column:game_id;default:0;comment:游戏ID;index:ix_game"`
-	AgentId         int             `json:"agent_id" gorm:"column:agent_id;default:0;comment:渠道ID"`
-	SiteId          int             `json:"site_id" gorm:"column:site_id;default:0;comment:广告位ID;index:ix_site"`
+	PlatformId      int64           `json:"platform_id" gorm:"column:platform_id;default:0;comment:平台ID"`
+	GameId          int64           `json:"game_id" gorm:"column:game_id;default:0;comment:游戏ID;index:ix_game"`
+	AgentId         int64           `json:"agent_id" gorm:"column:agent_id;default:0;comment:渠道ID"`
+	SiteId          int64           `json:"site_id" gorm:"column:site_id;default:0;comment:广告位ID;index:ix_site"`
 	Status          string          `json:"status" gorm:"size:50;column:status;default:'';comment:状态"`
 	GamePackagePath string          `json:"game_package_path" gorm:"size:512;column:game_package_path;default:'';comment:游戏包路径"`
 	ExecCmd         string          `json:"exec_cmd" gorm:"size:512;column:exec_cmd;default:'';comment:执行的命令"`
